test: move electron mocks
